Move serviceImpl to integration tests
